[debug] Move stack backtrace and symbol profiling to C library
