+++ elkscmd/debug/debug/instrument.h
@@ -1,0 +1,1 @@
+../../../libc/include/debug/instrument.h
