+++ elkscmd/debug/debug/syms.h
@@ -1,0 +1,1 @@
+../../../libc/include/debug/syms.h
